Refactor: 메모 화면 위젯 분리 #137
