+++ lib/features/memo/presentation/widgets/memo_empty_view.dart
@@ -1,0 +1,48 @@
+import 'package:commonplant_frontend/core/assets/app_icon_assets.dart';
+import 'package:commonplant_frontend/core/theme/app_colors.dart';
+import 'package:commonplant_frontend/core/theme/app_sizes.dart';
+import 'package:commonplant_frontend/core/theme/app_spacing.dart';
+import 'package:commonplant_frontend/core/theme/app_text_styles.dart';
+import 'package:commonplant_frontend/shared/widgets/common_svg_icon.dart';
+import 'package:flutter/material.dart';
+
+class MemoEmptyView extends StatelessWidget {
+  const MemoEmptyView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Padding(
+        padding: const EdgeInsets.symmetric(horizontal: AppSpacing.x20),
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            const CommonSvgIcon(
+              AppIconAssets.plantEmpty,
+              width: AppSizes.iconLarge,
+              height: AppSizes.iconLarge,
+              color: AppColors.iconInactive,
+              semanticsLabel: '메모 없음',
+            ),
+            const SizedBox(height: AppSpacing.x16),
+            Text(
+              '아직 작성된 메모가 없어요',
+              textAlign: TextAlign.center,
+              style: AppTextStyles.size18Medium.copyWith(
+                color: AppColors.textStrong,
+              ),
+            ),
+            const SizedBox(height: AppSpacing.x8),
+            Text(
+              '식물의 변화를 메모로 남겨보세요',
+              textAlign: TextAlign.center,
+              style: AppTextStyles.size14Medium.copyWith(
+                color: AppColors.textBody,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
